Refactor mixed-precision Fisher computation
